+++ pkg/types/stack.go
@@ -4,7 +4,7 @@ type StackResult struct {
 	Name           string    `json:"name"`
 	Components     []string  `json:"components"`
 	Configurations StackConf `json:"configurations,omitempty"`
-	Score          float64   `json:"score"`
+	Score          int       `json:"score"`
 }
 
 type Stack struct {
